Added basic styling for CSS
A link to the stylesheet was added for each recipe so that I could practice utilizing CSS via the external method. Each recipe recieved a special header title color and was centered along with the picture beneath it. The homepage was also updated to appear larger and centered via CSS.