Restyled the homepage with block model concepts.
After learning about the block model and utilizing paddings, borders, and margins, the homepage was updated to reflect what I have learned in a stylish and creative way.
The heading text of the homepage received a border and shadow-style text while the recipe cards were defined for a larger click area for the links and a basic appearence overhaul.
The image of the recipe was also incorporated onto the homepage and the colors have been streamlined for the homepage.